logging the values submitted in child form component within Parent chatbox component
