+++ frontend/src/components/Report.js
@@ -2,18 +2,7 @@ import React from 'react';
 import MarkdownView from 'react-showdown';
 
 export default function Report({data}) {
-    const copyToClipboard = () => {
-        const textarea = document.createElement('textarea');
-        textarea.id = 'temp_element';
-        textarea.style.height = 0;
-        document.body.appendChild(textarea);
-        textarea.value = document.getElementById('reportContainer')?.innerText;
-        const selector = document.querySelector('#temp_element');
-        selector.select();
-        document.execCommand('copy');
-        document.body.removeChild(textarea);
-    }
-
+    
     const markdown = `
         # Welcome to React Showdown :+1:
